Suggested changes
Let's get into some detail with this:

1. Dependencies go first, let's put them visible and distinguishable, a
good align of the equal signs might help to ease the reading of the
code.

2. Let's define a variable to hold all the critical behaviours and
shared information between files/modules, this variable can be `game`
(the word I'm using), it can be as simple as a literal object (`{}`).

3. Let's save in the game object all the global-classes, global methods
or non-constants (such as variables to store temporary or in memory
information). Here I'm storing the result of `express()`, the http
variable (just as an example, it's unlikely that you'll use it in
another file), the io variable (previously global) and the aux variable.

4. Pass the game variable to all your sub-modules (things that you
require that directly affect the behavior of your application, including
those submodules that define settings): here I'm just passing the game
variabel to the routes, but you can be more creative.

5. In the sub-modules, export a function that will receive the game
object. It's important to have a declaration of the game object outside
in case you want to put some other functions outside of the exported
function and if those functions will eventually use the game variable,
therefore you receive the game variable with a slight change of name
(_game) and you overwrite the one available to the whole file (`var
game;`).

6. Inside your exported function, and everywhere in that module's file,
you'll use global variables as elements of the game object, this way you
can store volatile information too: if any sub-module adds an element to
the game object, that element will be immediatly available to any other
submodule, or if they change a specific element of the game object, the
change is reflected everywhere in your server. I'm only using it to give
access to the `game.aux` and `game.io` objects which were global before
or were required two times, one in the index.js file and another time
per each the route.